pyproject.toml
[project]
name = "hibiri"
requires-python = ">= 3.10,<3.14"
description = "Simultaneous Speech-to-Speech Translation Without Aligned Data."
version = "0.0.4"
readme = "README.md"
dependencies = [
    # TODO: use the public release of moshi when it's been published
    "moshi>=0.2.13",
    "gradio>=5.50.0",
    "jupyter>=1.1.1",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "python-webrtcvad>=0.2.0",
    "librosa>=0.10.2",
    "websockets>=12.0",
]

[tool.setuptools.dynamic]
version = {attr = "hibiri.__version__"}

[tool.ruff]
line-length = 100

[tool.ruff.format]
line-ending = "lf"
# avoid having a trailing comma changing
# the behavior of the formatter
skip-magic-trailing-comma = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hibiri"]
artifacts = [
    "hibiri/samples/**",
    "hibiri/static/**",
]

[tool.hatch.metadata]
allow-direct-references = true

[project.scripts]
hibiri = "hibiri.run:cli_app"

[tool.uv]
# System python distribution can have header files missing, so we use only the python
# distributions provided by uv which have everything included.
python-preference = "only-managed"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
